refactor video components
